Chore: rename wrong name
오타 수정
